Serialise terminated instances
In order to retrieve information post-mortem (done, failed, lost) for an
instance we need to store it as the original data is lost with the
removal of the instance directory. This change-set serialises the
instance state into JSON as the value for the already existing keys in
the procs subpath. Together with that it implements a method to retrieve
the state and turns it into an `Instance` again.

This should be used in tooling (cli) to retrieve the list of all done instances.

* serialise and deserialise instances
* new Termination type
* retrieve methods on Proc for done/failed/lost