Introduce TrafficControl for procs
For some notion of control over how much traffic a proc should receive
this change introduces a new port per proc which can be used to
communicate the intended traffic share in percentage.

Along with that new unique port per proc comes another configurable
called TrafficControl. Which is meant to be `nil` if control should be
tuned off and the proc should get full traffic. When set it contains a
single field called Share and must be a positive integer between 0 and
100. The Share set will be communicate over some simple protocol via the
ControlPort.

* add ControlPort to proc
* add TrafficControl attribute to proc
* emit event when proc attrs change