Fix type of virtual instance statuses
The type of statuses failed, exited, lost and done was string so far,
instead of the expected InsStatus. In most cases this doesn't matter,
e.g. when the constants are used as untyped constants and usable as both
strings as well as InsStatus.

For more information on untyped constants in golang, this blog post
might be a good start: http://blog.golang.org/constants