Fix emission of early EvInsReg
Up until now the crucial registration event for instances was fired when
a Set on /instances/$id/object was observed. As this is not the last
file written for an instance there is a window where the event is
observed and the event tries to provide a complete Instance object from
the partially written instance sub-tree.

To avoid this error case we listen now on the last path written and emit
when we can be sure that the entire instance sub-tree is present. This
relies on a stable order of file creation during RegisterInstance.