Fix WatchEvent for old events
This fixes a bug during event enrichment for an instance which has been
unregistered in the meantime. As some clients might lack behind in the
timeline and process events for old instances, it's important that the
Snapshot gets never been forwarded during event enrichment. A test has
been added to ensure the correct behavior.