Fix WaitFailed to set all changed attributes
So far (Instance) WaitFailed() only updated the Status attribute, but
not the Termination attribute. As the latter includes the actually
interesting reason for a failure, this change updates that information
as well now. In order to do that, the Wait method had to be changed to
wait for all path updates exercised by Failed().